convert most floating types to single-precision
fixed #114 -- decided that it's unnecessary to support both types right now, although might be useful to typedef all the types in the API for clarity

+++ examples/common/wav_file.cpp
@@ -9,7 +9,7 @@
 
 namespace barely::examples {
 
-const std::vector<double>& WavFile::GetData() const noexcept { return data_; }
+const std::vector<float>& WavFile::GetData() const noexcept { return data_; }
 
 int WavFile::GetChannelCount() const noexcept { return channel_count_; }
