+++ platforms/unity/Assets/BarelyMusician/Scripts/GlobalTempo.cs
@@ -5,10 +5,10 @@ namespace Barely {
   public class GlobalTempo : MonoBehaviour {
     [Tooltip("Sets the global tempo in beats per minute.")]
     [Range(30.0f, 960.0f)]
-    public double Bpm = 120.0;
+    public float Bpm = 120.0f;
 
     private void Update() {
-      Musician.Tempo = Bpm;
+      Musician.Tempo = (double)Bpm;
     }
   }
 }  // namespace Barely
